+++ packages/graphql-language-service-server/CHANGELOG.md
@@ -1,5 +1,18 @@
 # graphql-language-service-server
 
+## 2.7.16
+
+### Patch Changes
+
+- Updated dependencies [[`2502a364`](https://github.com/graphql/graphiql/commit/2502a364b74dc754d92baa1579b536cf42139958)]:
+  - graphql-language-service@5.0.0
+
+## 2.7.15
+
+### Patch Changes
+
+- [#2214](https://github.com/graphql/graphiql/pull/2214) [`ab83198f`](https://github.com/graphql/graphiql/commit/ab83198fa8b3c5453d3733982ee9ca8a2d6bca7a) Thanks [@Cellule](https://github.com/Cellule)! - Fixed Windows fileUri when resolving type definition location
+
 ## 2.7.14
 
 ### Patch Changes